turn checkl into a repo viewer
