+++ package.json
@@ -6,15 +6,19 @@
   "dependencies": {
     "@akebifiky/remark-simple-plantuml": "^1.0.2",
     "@rjsf/core": "^3.0.0",
+    "@tanstack/react-query": "4.15.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "gh-pages": "^3.2.3",
     "handlebars": "^4.7.7",
+    "highlight.js": "^11.6.0",
     "mermaid": "^8.11.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-markdown": "^6.0.2",
+    "react-router": "^6.4.3",
+    "react-router-dom": "^6.4.3",
     "react-scripts": "4.0.3",
     "remark-autolink-headings": "^6.0.1",
     "remark-emoji": "^2.2.0",
