+++ public/index.html
@@ -22,9 +22,11 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/github-fork-ribbon-css/0.2.3/gh-fork-ribbon.min.css" />
+    <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/11.6.0/styles/default.min.css">
     <title>Checkl - your dynamic checklist</title>
   </head>
   <body >
+    <script src="//cdnjs.cloudflare.com/ajax/libs/highlight.js/11.6.0/highlight.min.js"></script>    
     <style>
       @media print {
         @media print {
@@ -72,6 +74,5 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
-    <a class="github-fork-ribbon noprint" target="_blank" href="https://github.com/mestachs/checkl/blob/master/README.md" data-ribbon="Documentation" title="I'm on github">Documentation</a>
-  </body>
+   </body>
 </html>
